+++ lib/WriteOff/Schema/ResultSet.pm
@@ -21,9 +21,9 @@ sub now {
 }
 
 sub now_dt {
-	return DateTime->now;
+	#return DateTime->now;
 	
-	return shift->parse_datetime('2012-11-29 02:00:00');
+	return shift->parse_datetime('2012-12-03 02:00:00');
 }
 
 sub created_before {